refactor: redux store 로직 분리
